+++ src/lib.rs
@@ -5,8 +5,6 @@ mod fluentd;
 mod matcher;
 mod producer;
 mod report;
-#[allow(dead_code)]
-mod session;
 
 pub use config::{Config, InputType, OutputType};
 pub use controller::Controller;